Authorization

+++ client/src/components/WorkoutDtails.jsx
@@ -1,14 +1,22 @@
 import React from 'react'
 import {useWorkoutsContext} from '../hooks/useWorkoutsContext'
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import { useAuthContext } from '../hooks/useAuthContext'
 import Swal from 'sweetalert2'
 
 function WorkoutDtails({workout}) {
   const {dispatch} = useWorkoutsContext();
+  const {user} = useAuthContext();
   const handleClick = async () =>{
+    if(!user){
+      return
+    }
 
     const response = await fetch ('http://localhost:4000/api/workouts/'+ workout._id, {
-      method : 'DELETE'
+      method : 'DELETE',
+      headers :{
+        'Authorization' : `Bearer ${user.token}`
+      }
     })
     const Toast = Swal.mixin({
       toast: true,
